abort reinitialization
initializing again might break --toggle.
When overlayfs is enabled and a reinitialization is attempted,
cmdline.txt.orig will be overwritten with the current cmdline.txt.overlay.
(because cmdline.txt equals cmdline.txt.overlay at that point and will
be backed up).
Additionally 'boot=overlay' will be added a second time to cmdline.txt.
Overall it does not seem to be a good idea to reinitialize, therefore
such an attempt should be aborted.

Signed-off-by: Julian Heuking <jheuking@googlemail.com>